frontend: Dark Mode - Allowing Overrides
