+++ frontend/packages/core/src/Theme/types.tsx
@@ -1,11 +1,9 @@
-import type { Color } from "@mui/material";
+import type { Color, Theme as MuiTheme } from "@mui/material";
+import type { PaletteOptions as MuiPaletteOptions } from "@mui/material/styles";
 
-export type ThemeVariant = "light" | "dark";
+import type VARIANTS from "./variants";
 
-export enum THEME_VARIANTS {
-  light = "light",
-  dark = "dark",
-}
+export type ThemeVariant = keyof typeof VARIANTS;
 
 export interface StrokeColor {
   primary: string;
@@ -14,21 +12,13 @@ export interface StrokeColor {
   inverse: string;
 }
 
-export interface ChartColors {
-  common: {
-    data: string[];
-  };
-  pie: {
-    labelPrimary: string;
-    labelSecondary: string;
-  };
-  linearTimeline: {
-    xAxisStroke: string;
-    tooltipBackgroundColor: string;
-    tooltipTextColor: string;
-    gridBackgroundColor: string;
-    gridStroke: string;
-  };
+export interface ThemeOptions extends MuiTheme {}
+
+export interface PaletteOptions extends MuiPaletteOptions {
+  type: ThemeVariant;
+  contrastColor: string;
+  headerGradient: string;
+  brandColor: string;
 }
 
 export interface ClutchColors {
@@ -37,7 +27,62 @@ export interface ClutchColors {
   green: Color;
   amber: Color;
   red: Color;
-  charts: ChartColors;
 }
+
+interface CommonChartColors {
+  data: string[];
+}
+
+interface PieChartColors {
+  labelPrimary: string;
+  labelSecondary: string;
+}
+
+interface TimelineChartColors {
+  xAxisStroke: string;
+  tooltipBackgroundColor: string;
+  tooltipTextColor: string;
+  gridBackgroundColor: string;
+  gridStroke: string;
+}
+
+export interface ChartColors {
+  common: CommonChartColors;
+  pie: PieChartColors;
+  linearTimeline: TimelineChartColors;
+}
+
+export interface ClutchTheme {
+  colors: ClutchColors;
+  chartColors: ChartColors;
+}
+
+type Partial<T> = {
+  [P in keyof T]?: T[P];
+};
+
+export type CustomClutchColor = Partial<Color>;
+export type CustomClutchColors = {
+  [P in keyof ClutchColors]?: CustomClutchColor;
+};
+export type CustomChartColors = {
+  [P in keyof ChartColors]?: Partial<ChartColors[P]>;
+};
+export type CustomPalette = {
+  [P in keyof PaletteOptions]?: Partial<PaletteOptions[P]>;
+};
+export type CustomThemeOptions = {
+  [P in keyof ThemeOptions]?: Partial<ThemeOptions[P]>;
+};
+export interface CustomClutchTheme {
+  colors?: CustomClutchColors;
+  chartColors?: CustomChartColors;
+  palette?: CustomPalette;
+  themeOptions?: CustomThemeOptions;
+}
+
+export type ThemeOverrides = {
+  [P in ThemeVariant]?: CustomClutchTheme;
+};
 
 export type ComponentState = "hover" | "focused" | "pressed" | "selected" | "disabled";
