+++ frontend/packages/core/src/Theme/variants/dark/colors.tsx
@@ -1,0 +1,98 @@
+import type { Color } from "@mui/material";
+
+import type { ClutchColors } from "../../types";
+
+const neutral: Color = {
+  50: "#232542",
+  100: "#30324E",
+  200: "#3E4059",
+  300: "#55566D",
+  400: "#77788A",
+  500: "#8D8E9E",
+  600: "#A4A5B1",
+  700: "#D2D2D8",
+  800: "#E8E8EB",
+  900: "#FFFFFF",
+  A100: "#FFFFFF",
+  A200: "#FFFFFF",
+  A400: "#FFFFFF",
+  A700: "#FFFFFF",
+};
+
+const blue: Color = {
+  50: "#090A42",
+  100: "#050656",
+  200: "#0A086B",
+  300: "#13199D",
+  400: "#2A4FF6",
+  500: "#4281F6",
+  600: "#5AABF6",
+  700: "#8CC4F8",
+  800: "#C2E1FE",
+  900: "#DCECFB",
+  A100: "#DCECFB",
+  A200: "#DCECFB",
+  A400: "#DCECFB",
+  A700: "#DCECFB",
+};
+
+const green: Color = {
+  50: "#002C05",
+  100: "#084713",
+  200: "#145F1D",
+  300: "#217A2A",
+  400: "#2D9638",
+  500: "#54B45B",
+  600: "#73C178",
+  700: "#9CD29E",
+  800: "#C3E4C4",
+  900: "#E6F4E7",
+  A100: "#E6F4E7",
+  A200: "#E6F4E7",
+  A400: "#E6F4E7",
+  A700: "#E6F4E7",
+};
+
+const amber: Color = {
+  50: "#352215",
+  100: "#7E3E02",
+  200: "#985304",
+  300: "#B26A08",
+  400: "#CC820D",
+  500: "#E69F2A",
+  600: "#EAB04E",
+  700: "#EFC67F",
+  800: "#F6DCB1",
+  900: "#FBF1E0",
+  A100: "#FBF1E0",
+  A200: "#FBF1E0",
+  A400: "#FBF1E0",
+  A700: "#FBF1E0",
+};
+
+const red: Color = {
+  50: "#501306",
+  100: "#621809",
+  200: "#751E0D",
+  300: "#972814",
+  400: "#B3351E",
+  500: "#C8482C",
+  600: "#EB7A60",
+  700: "#F4A08A",
+  800: "#F0B9AB",
+  900: "#FBE8E7",
+  A100: "#FFFFFF",
+  A200: "#FFFFFF",
+  A400: "#FFFFFF",
+  A700: "#FFFFFF",
+};
+
+const colors: ClutchColors = {
+  neutral,
+  blue,
+  green,
+  amber,
+  red,
+};
+
+export default colors;
