+++ frontend/packages/core/src/Theme/variants/dark/index.tsx
@@ -1,0 +1,11 @@
+import type { ClutchTheme } from "../../types";
+
+import chartColors from "./chart";
+import colors from "./colors";
+
+const DARK_THEME: ClutchTheme = {
+  chartColors,
+  colors,
+};
+
+export default DARK_THEME;
