+++ frontend/packages/core/src/Theme/variants/index.tsx
@@ -1,0 +1,9 @@
+import DARK_THEME from "./dark";
+import LIGHT_THEME from "./light";
+
+const VARIANTS = {
+  DARK: DARK_THEME,
+  LIGHT: LIGHT_THEME,
+};
+
+export default VARIANTS;
