+++ frontend/packages/core/src/Theme/variants/light/colors.tsx
@@ -1,0 +1,98 @@
+import type { Color } from "@mui/material";
+
+import type { ClutchColors } from "../../types";
+
+const neutral: Color = {
+  50: "#F8F8F9",
+  100: "#F0F1F3",
+  200: "#DBDBE0",
+  300: "#C2C3CB",
+  400: "#9E9FAC",
+  500: "#868798",
+  600: "#56586E",
+  700: "#3D4059",
+  800: "#252845",
+  900: "#0D1030",
+  A100: "#0D1030",
+  A200: "#0D1030",
+  A400: "#0D1030",
+  A700: "#0D1030",
+};
+
+const blue: Color = {
+  50: "#F9F9FE",
+  100: "#F5F6FD",
+  200: "#EBEDFB",
+  300: "#D7DAF6",
+  400: "#C2C8F2",
+  500: "#727FE1",
+  600: "#3548D4",
+  700: "#1629B9",
+  800: "#0A1CA6",
+  900: "#011082",
+  A100: "#011082",
+  A200: "#011082",
+  A400: "#011082",
+  A700: "#011082",
+};
+
+const green: Color = {
+  50: "#E5FCE8",
+  100: "#CBF7CF",
+  200: "#ACF2B2",
+  300: "#6CD47A",
+  400: "#4AB958",
+  500: "#32A140",
+  600: "#1C872A",
+  700: "#106E1D",
+  800: "#086515",
+  900: "#02590E",
+  A100: "#02590E",
+  A200: "#02590E",
+  A400: "#02590E",
+  A700: "#02590E",
+};
+
+const amber: Color = {
+  50: "#FFFBEB",
+  100: "#FEF3C7",
+  200: "#FDE68A",
+  300: "#FCD34D",
+  400: "#FBBF24",
+  500: "#F59E0B",
+  600: "#D97706",
+  700: "#B45309",
+  800: "#92400E",
+  900: "#78350F",
+  A100: "#78350F",
+  A200: "#78350F",
+  A400: "#78350F",
+  A700: "#78350F",
+};
+
+const red: Color = {
+  50: "#FFF4F3",
+  100: "#FDDCD2",
+  200: "#F6A996",
+  300: "#F3886E",
+  400: "#F26E50",
+  500: "#F15534",
+  600: "#CA4428",
+  700: "#A1301C",
+  800: "#792111",
+  900: "#571608",
+  A100: "#571608",
+  A200: "#571608",
+  A400: "#571608",
+  A700: "#571608",
+};
+
+const colors: ClutchColors = {
+  neutral,
+  blue,
+  green,
+  amber,
+  red,
+};
+
+export default colors;
