+++ frontend/packages/core/src/Theme/variants/light/index.tsx
@@ -1,0 +1,11 @@
+import type { ClutchTheme } from "../../types";
+
+import chartColors from "./chart";
+import colors from "./colors";
+
+const LIGHT: ClutchTheme = {
+  chartColors,
+  colors,
+};
+
+export default LIGHT;
